+++ run_tests.sh
@@ -1,11 +1,20 @@
 #!/usr/bin/env bash
 
-images=("php" "fpm" "apache-php" "akeneo-fpm")
-versions=("7.1" "7.1" "php-7.1" "php-7.1")
+didFail=0
+images=("php" "fpm" "apache-php" )
+versions=("7.1" "php-7.1" "php-7.1")
 
 cwd=$(pwd)
 for (( i=0; i<${#images[@]}; i++ ));
 do
     cd ${cwd}/${images[i]}
     docker build -t akeneo/${images[i]}:${versions[i]} .
+
+    if ! docker inspect akeneo/${images[i]}:${versions[i]} &> /dev/null; then
+        echo $'\timage does not exist!'
+        didFail=1
+        continue
+    fi
 done
+
+exit ${didFail}